requirements.txt
python-dotenv
langchain
pypdf
tiktoken
faiss-cpu
chroma
PyPDF2
transformers
gradio
sentence_transformers
text_generation
llama_index
torch


diffusers
bitsandbytes==0.40.2
cuda-python